+++ src/pages/result/index.style.tsx
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const Container = styled.div`
+  width: calc(var(--vw, 1vw) * 100);
+  height: calc(var(--vh, 1vh) * 100);
+`;
