+++ src/pages/result/index.tsx
@@ -1,34 +1,16 @@
 import { useLayoutEffect, useState } from "react";
 import { useLocation, useNavigate } from "react-router-dom";
-import {
-  BottomBar,
-  CoverPage,
-  DefinitionPage,
-  LifePage,
-  MedicinePage,
-  TreatmentPage,
-} from "../components/resultPage";
-import ResultHeader from "../components/header/ResultHeader";
-import ModalContainer from "../components/modal/ModalContainer";
-import ResultModal from "../components/modal/ResultModal";
+import { BottomBar, CoverPage, DefinitionPage, LifePage, MedicinePage, TreatmentPage } from "../../components/resultPage";
+import ResultHeader from "../../components/header/ResultHeader";
+import ModalContainer from "../../components/modal/ModalContainer";
+import ResultModal from "../../components/modal/ResultModal";
 
 import { Swiper, SwiperSlide } from "swiper/react";
 import "swiper/css";
-import ResultLoading from "../components/loading/ResultLoading";
-import { IDiagnosticResult } from "../interfaces/diagnosticResult";
-import {
-  ICoverPageProps,
-  IDefinePageProps,
-  ILifeProps,
-  IMedicine,
-  ITreatPageProps,
-} from "../interfaces/resultPage";
-import styled from "styled-components";
-
-const Container = styled.div`
-  width: calc(var(--vw, 1vw) * 100);
-  height: calc(var(--vh, 1vh) * 100);
-`;
+import ResultLoading from "../../components/loading/ResultLoading";
+import { IDiagnosticResult } from "../../interfaces/diagnosticResult";
+import { ICoverPageProps, IDefinePageProps, ILifeProps, IMedicine, ITreatPageProps } from "../../interfaces/resultPage";
+import { Container } from "./index.style";
 
 const ResultPage = () => {
   const navigate = useNavigate();
@@ -112,10 +94,7 @@ const ResultPage = () => {
               <DefinitionPage defineData={defineData} />
             </SwiperSlide>
             <SwiperSlide>
-              <LifePage
-                lifestyle={lifeData}
-                type={state.diagnostic_result.id}
-              />
+              <LifePage lifestyle={lifeData} type={state.diagnostic_result.id} />
             </SwiperSlide>
             {medicineData && (
               <SwiperSlide>
@@ -130,11 +109,7 @@ const ResultPage = () => {
           </Swiper>
           {modal && (
             <ModalContainer setModal={setModal}>
-              <ResultModal
-                setModal={setModal}
-                setLoading={setLoading}
-                resultId={state.diagnostic_result.id}
-              />
+              <ResultModal setModal={setModal} setLoading={setLoading} resultId={state.diagnostic_result.id} />
             </ModalContainer>
           )}
           <BottomBar
