+++ src/pages/symptom/index.tsx
@@ -1,33 +1,11 @@
 import { Suspense, useState, useEffect } from "react";
 import { Canvas } from "@react-three/fiber";
-import styled from "styled-components";
-import Character from "../components/symptomPage/Character";
-import Overlay from "../components/symptomPage/Overlay";
+import Character from "../../components/symptomPage/Character";
+import Overlay from "../../components/symptomPage/Overlay";
 import { useLocation, useNavigate } from "react-router-dom";
-import theme from "../lib/theme";
-import ContentHeader from "../components/header/ContentHeader";
-
-const Container = styled.section`
-  width: calc(var(--vw, 1vw) * 100);
-  height: calc(var(--vh, 1vh) * 100);
-
-  background: #0c0d10;
-
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-`;
-
-const CanvasSection = styled.div`
-  width: calc(var(--vw, 1vw) * 100);
-  height: calc(var(--vh, 1vh) * 100);
-
-  z-index: 0;
-  position: absolute;
-
-  background: ${({ theme }) => theme.color.grey_900};
-`;
+import theme from "../../lib/theme";
+import ContentHeader from "../../components/header/ContentHeader";
+import { Container, CanvasSection } from "./index.style";
 
 const Point = ({ x, y, z }: { x: number; y: number; z: number }) => {
   return (
